requirements.txt
# Core dependencies
streamlit==1.29.0
openai==1.6.1
chromadb==0.4.22
pypdf==3.17.4
python-dotenv==1.0.0

# Utilities
tiktoken==0.5.2
numpy==1.24.3
